+++ dsh_sdk/CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2024-07-01
+### Added
+- Add token fetcher for DSH REST API
+- Add platform enum with metadata
+- Add example on how to use the token fetcher and rest api client
+
+
+## [0.4.1] - 2024-06-13
+### Added
+- Add loading Certificates and Keys from $PKI_CONFIG_DIR
+  - Compatbile with DSH VPN
+  - Compatbile with Kafka Proxy
+- Overwrrite tenant name via $DSH_TENANT_NAME variable
+
+### Changed
+- Restructure of the private functions to make it more modular
+- Improved logging
+- Improved API Documentation
+
 ## [0.4.0] - 2024-04-25
 
 ### Fixed